streamlit-fastapi-service/Backend/requirements.txt
apache-airflow
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
asyncio
websockets
snowflake
snowflake-connector-python
pinecone-client
openai
boto3
pyyaml
